- Renamed `ReadEvents` to `EventReader` - Read events using Java tests using `EventReader`
